compatibility with newer OS

+++ src/leveldb/db/filename.cc
@@ -4,7 +4,6 @@
 
 #include <ctype.h>
 #include <stdio.h>
-#include <stdint.h>
 #include "db/filename.h"
 #include "db/dbformat.h"
 #include "leveldb/env.h"
@@ -19,8 +18,8 @@ extern Status WriteStringToFileSync(Env* env, const Slice& data,
 static std::string MakeFileName(const std::string& name, uint64_t number,
                                 const char* suffix) {
   char buf[100];
-  snprintf(buf, sizeof(buf), "/%06" PRIu64 ".%s",
-           static_cast<uint64_t>(number),
+  snprintf(buf, sizeof(buf), "/%06llu.%s",
+           static_cast<unsigned long long>(number),
            suffix);
   return name + buf;
 }
@@ -43,8 +42,8 @@ std::string SSTTableFileName(const std::string& name, uint64_t number) {
 std::string DescriptorFileName(const std::string& dbname, uint64_t number) {
   assert(number > 0);
   char buf[100];
-  snprintf(buf, sizeof(buf), "/MANIFEST-%06" PRIu64,
-           static_cast<uint64_t>(number));
+  snprintf(buf, sizeof(buf), "/MANIFEST-%06llu",
+           static_cast<unsigned long long>(number));
   return dbname + buf;
 }
